build: split up building of libs, bins, modules, benchmark
this is needed as a preparation for more build refactoring. Overall goal
to build efl in a single .so.

What this commit results in is that *every* lib gets build first, then
the modules and binaries of all subprojects. This way of building is
needed in order to replace the split up libraries with the efl-one one
later.

Additionally, the infrastructure is added to set flags on the libraries
that should be build together.

Differential Revision: https://phab.enlightenment.org/D11805